server/requirements.txt
flask
peewee
werkzeug
pylint
python-dotenv
Flask-CORS
Flask-Bcrypt
flask-login
waitress
nanoid